Move CDC_Erite test to slow
