+++ ydb/core/persqueue/ut/slow/ya.make
@@ -12,11 +12,17 @@ PEERDIR(
     library/cpp/svnversion
     ydb/core/persqueue/ut/common
     ydb/core/testlib/default
+    ydb/public/sdk/cpp/client/ydb_persqueue_core/ut/ut_utils
+    ydb/public/sdk/cpp/client/ydb_persqueue_public/ut/ut_utils
+    ydb/public/sdk/cpp/client/ydb_topic/ut/ut_utils
+
+    ydb/core/tx/schemeshard/ut_helpers
 )
 
 YQL_LAST_ABI_VERSION()
 
 SRCS(
+    autopartitioning_ut.cpp
     pq_ut.cpp
 )
